// src/modules/chat/enums/chat-events.enums.ts
export enum ChatEvents {
  joinRoom = 'joinRoom',
  roomJoined = 'roomJoined',
  leaveRoom = 'leaveRoom',
  roomLeft = 'roomLeft',
  sendMessage = 'sendMessage',
  messageSent = 'messageSent',
  roomUpdated = 'roomUpdated',
  exception = 'exception',
  disconnect = 'disconnect',
  connect = 'connect',
}